Make cisagov/ansible-role-freeipa-client a role dependency
The FreeIPA client software must be installed in order for the joiner
user to perform its function anyway.  Previously we did not list
cisagov/ansible-role-freeipa-client as a role dependency and we just
installed the FreeIPA client package via the Molecule prepare
playbook, but this now causes an error to be issued by ansible-lint.